1. Mob will deal the damaged that only specified in DB

+++ src/main/java/com/phazerous/phazerous/entities/models/entities/MobEntity.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 public class MobEntity extends BaseEntity {
     private Integer mobType;
     private Long maxHealth;
-    private Double attack;
+    private Long attack;
     private MoneyReward moneyReward;
 }
